[devel-interpolators-minor-1]
1. Renamed
    a. `GaussSeidel.py` -> `gauss_seidel.py`
    b. `GaussJacobi.py` -> `gauss_jacobi.py`
    c. `ITERATIVE_SOLVER_.py` -> `iterative_solvers_.py`
    d. `newton_cotes.py` -> `newton_cotes_.py`
    e. `PREDICTOR_CORRECTOR.py` -> `predictor_corrector_.py`
    f. `PredictorCorrectorMethods.py` -> `predictor_corrector_methods.py`
2. Minor changes
