+++ src/Solvers/__backend/newton_cotes_.py
